I2C retries and some refinements

+++ ROS2/ros2_scara_ws/src/joint_communication/include/joint_communication/mJointCom.h
@@ -22,7 +22,7 @@ public:
  * @param offset offset between encoder zero and joint zero (in joint units)
 
  */
-  void addJoint(const int address, const std::string name, const int gearRatio, const int offset);
+  void addJoint(const int address, const std::string name, const float gearRatio, const int offset);
 
   /**
  * Initializes the drivers
@@ -120,7 +120,7 @@ public:
   /**
    * @brief Enable TMC5130 StallGuards for each driver
    * The threshold should be tuned as to trigger stallguard before a step is lost.
-   * @param threshold stall sensitivity. A value between -64 and +63
+   * @param threshold stall sensitivity. A value between -64 and +63, low is more sensitive
    */
   int enableStallguards(std::vector<int8_t> thresholds);
